requirements.txt
asgiref==3.7.2
Django==4.2.7
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
Markdown==3.5.1
pytz==2023.3.post1
setuptools==69.0.2
sqlparse==0.4.4
wheel==0.42.0
